refactor: split runtime handlers
